Refactoring everything in `main.tsx` to separate files
